Add files via upload
This dataset represents the Biochemical Oxygen Demand(BOD) in river water in 7 different stations. Dataset contains 9 features with 147 observations. Most of the observation values are empty. There are 569 missing values in total. There is a target column named 'target'.  Most of the observations in columns 3,4,5,6 and 7 are missing and thus these columns are removed from the data for further analysis. Columns 1 and 2 have two missing values which are removed. In this project different models are used to compare the accuracy of output from each model. Models such as Linear Regression, Decision Tree, Random Forest and XGB Regressor are used. The accuracy of outputs from each model is compared using MAE, MSE, R2 and cross validation. Using all these it was concluded that XGB Regressor gave better accuracy than all other models.